feat: 유저 정보(총 KM, 여행 개수) 요청 로직 구현 [TRL-72]
